+++ frontend-plugin-core/src/main/java/com/github/eirslett/maven/plugins/frontend/lib/NodeInstaller.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.List;
 
 import org.apache.commons.io.FileUtils;
 import org.slf4j.Logger;
@@ -21,15 +22,15 @@ public class NodeInstaller {
 
     private final Logger logger;
 
-    private final InstallConfig config;
+    private final List<InstallConfig> configs;
 
     private final ArchiveExtractor archiveExtractor;
 
     private final FileDownloader fileDownloader;
 
-    NodeInstaller(InstallConfig config, ArchiveExtractor archiveExtractor, FileDownloader fileDownloader) {
+    NodeInstaller(List<InstallConfig> configs, ArchiveExtractor archiveExtractor, FileDownloader fileDownloader) {
         this.logger = LoggerFactory.getLogger(getClass());
-        this.config = config;
+        this.configs = configs;
         this.archiveExtractor = archiveExtractor;
         this.fileDownloader = fileDownloader;
     }
@@ -63,8 +64,9 @@ public class NodeInstaller {
         if (this.npmVersion != null) {
             if ("provided".equals(this.npmVersion)) {
                 if (Integer.parseInt(this.nodeVersion.replace("v", "").split("[.]")[0]) < 4) {
-                    throw new InstallationException("NPM version is '" + this.npmVersion
-                        + "' but Node didn't include NPM prior to v4.0.0");
+                    throw new InstallationException("NPM version is '"
+                            + this.npmVersion
+                            + "' but Node didn't include NPM prior to v4.0.0");
                 }
                 return true;
             }
@@ -79,38 +81,47 @@ public class NodeInstaller {
             if (this.nodeDownloadRoot == null || this.nodeDownloadRoot.isEmpty()) {
                 this.nodeDownloadRoot = DEFAULT_NODEJS_DOWNLOAD_ROOT;
             }
-            if (!nodeIsAlreadyInstalled()) {
-                this.logger.info("Installing node version {}", this.nodeVersion);
-                if (!this.nodeVersion.startsWith("v")) {
-                    this.logger.warn("Node version does not start with naming convention 'v'.");
-                }
-                if (this.config.getPlatform().isWindows()) {
-                    if (npmProvided()) {
-                        installNodeWithNpmForWindows();
-                    } else {
-                        installNodeForWindows();
-                    }
-                } else {
-                    installNodeDefault();
-                }
+            //Execution platform always 1st.
+            InstallConfig execConfig = this.configs.get(0);
+            for (int i = 0; i < this.configs.size(); ++i) {
+                processConfig(execConfig, this.configs.get(i));
             }
         }
     }
 
-    private boolean nodeIsAlreadyInstalled() {
+    private void processConfig(InstallConfig execConfig, InstallConfig config) throws InstallationException {
+        if(!nodeIsAlreadyInstalled(config)) {
+            this.logger.info("Installing node version {}", this.nodeVersion);
+            if (!this.nodeVersion.startsWith("v")) {
+                this.logger.warn("Node version does not start with naming convention 'v'.");
+            }
+            if (config.getPlatform().isWindows()) {
+                if (npmProvided()) {
+                    installNodeWithNpmForWindows(execConfig, config);
+                } else {
+                    installNodeForWindows(execConfig, config);
+                }
+            } else {
+                installNodeDefault(execConfig, config);
+            }
+        }
+    }
+
+    private boolean nodeIsAlreadyInstalled(InstallConfig execConfig) {
         try {
-            NodeExecutorConfig executorConfig = new InstallNodeExecutorConfig(this.config);
+            NodeExecutorConfig executorConfig = new InstallNodeExecutorConfig(execConfig);
             File nodeFile = executorConfig.getNodePath();
             if (nodeFile.exists()) {
-                final String version =
-                    new NodeExecutor(executorConfig, Arrays.asList("--version"), null).executeAndGetResult(logger);
+                final String version = new NodeExecutor(executorConfig, Arrays.asList("--version"), null).executeAndGetResult(logger);
 
                 if (version.equals(this.nodeVersion)) {
                     this.logger.info("Node {} is already installed.", version);
                     return true;
                 } else {
-                    this.logger.info("Node {} was installed, but we need version {}", version,
-                        this.nodeVersion);
+                    this.logger.info(
+                            "Node {} was installed, but we need version {}",
+                            version,
+                            this.nodeVersion);
                     return false;
                 }
             } else {
@@ -121,49 +132,55 @@ public class NodeInstaller {
         }
     }
 
-    private void installNodeDefault() throws InstallationException {
+    private void installNodeDefault(InstallConfig execConfig, InstallConfig config) throws InstallationException {
         try {
-            final String longNodeFilename =
-                this.config.getPlatform().getLongNodeFilename(this.nodeVersion, false);
+            final String longNodeFilename = config.getPlatform().getLongNodeFilename(this.nodeVersion, false);
             String downloadUrl = this.nodeDownloadRoot
-                + this.config.getPlatform().getNodeDownloadFilename(this.nodeVersion, false);
-            String classifier = this.config.getPlatform().getNodeClassifier();
+                    + config.getPlatform().getNodeDownloadFilename(this.nodeVersion, false);
+            String classifier = config.getPlatform().getNodeClassifier();
 
-            File tmpDirectory = getTempDirectory();
+            File tmpDirectory = getTempDirectory(execConfig);
 
-            CacheDescriptor cacheDescriptor = new CacheDescriptor("node", this.nodeVersion, classifier,
-                this.config.getPlatform().getArchiveExtension());
+            CacheDescriptor cacheDescriptor = new CacheDescriptor("node",
+                    this.nodeVersion,
+                    classifier,
+                    config.getPlatform().getArchiveExtension());
 
-            File archive = this.config.getCacheResolver().resolve(cacheDescriptor);
+            File archive = config.getCacheResolver().resolve(cacheDescriptor);
 
             downloadFileIfMissing(downloadUrl, archive, this.userName, this.password);
 
             extractFile(archive, tmpDirectory);
 
             // Search for the node binary
-            File nodeBinary =
-                new File(tmpDirectory, longNodeFilename + File.separator + "bin" + File.separator + "node");
+            File nodeBinary = new File(tmpDirectory, longNodeFilename + File.separator + "bin" + File.separator + "node");
             if (!nodeBinary.exists()) {
                 throw new FileNotFoundException(
-                    "Could not find the downloaded Node.js binary in " + nodeBinary);
+                        "Could not find the downloaded Node.js binary in " + nodeBinary);
             } else {
-                File destinationDirectory = getInstallDirectory();
+                File destinationDirectory = getInstallDirectory(execConfig);
 
                 File destination = new File(destinationDirectory, "node");
+                if(destination.exists()) {
+                    //Building multiple times on windows with file without extention will not be able to rename.
+                    destination.delete();
+                }
                 this.logger.info("Copying node binary from {} to {}", nodeBinary, destination);
                 if (!nodeBinary.renameTo(destination)) {
                     throw new InstallationException("Could not install Node: Was not allowed to rename "
-                        + nodeBinary + " to " + destination);
+                            + nodeBinary
+                            + " to "
+                            + destination);
                 }
 
                 if (!destination.setExecutable(true, false)) {
                     throw new InstallationException(
-                        "Could not install Node: Was not allowed to make " + destination + " executable.");
+                            "Could not install Node: Was not allowed to make " + destination + " executable.");
                 }
 
                 if (npmProvided()) {
                     File tmpNodeModulesDir = new File(tmpDirectory,
-                        longNodeFilename + File.separator + "lib" + File.separator + "node_modules");
+                            longNodeFilename + File.separator + "lib" + File.separator + "node_modules");
                     File nodeModulesDirectory = new File(destinationDirectory, "node_modules");
                     File npmDirectory = new File(nodeModulesDirectory, "npm");
                     FileUtils.copyDirectory(tmpNodeModulesDir, nodeModulesDirectory);
@@ -190,20 +207,21 @@ public class NodeInstaller {
         }
     }
 
-    private void installNodeWithNpmForWindows() throws InstallationException {
+    private void installNodeWithNpmForWindows(InstallConfig execConfig, InstallConfig config) throws InstallationException {
         try {
-            final String longNodeFilename =
-                this.config.getPlatform().getLongNodeFilename(this.nodeVersion, true);
+            final String longNodeFilename = config.getPlatform().getLongNodeFilename(this.nodeVersion, true);
             String downloadUrl = this.nodeDownloadRoot
-                + this.config.getPlatform().getNodeDownloadFilename(this.nodeVersion, true);
-            String classifier = this.config.getPlatform().getNodeClassifier();
+                    + config.getPlatform().getNodeDownloadFilename(this.nodeVersion, true);
+            String classifier = config.getPlatform().getNodeClassifier();
 
-            File tmpDirectory = getTempDirectory();
+            File tmpDirectory = getTempDirectory(config);
 
-            CacheDescriptor cacheDescriptor = new CacheDescriptor("node", this.nodeVersion, classifier,
-                this.config.getPlatform().getArchiveExtension());
+            CacheDescriptor cacheDescriptor = new CacheDescriptor("node",
+                    this.nodeVersion,
+                    classifier,
+                    config.getPlatform().getArchiveExtension());
 
-            File archive = this.config.getCacheResolver().resolve(cacheDescriptor);
+            File archive = config.getCacheResolver().resolve(cacheDescriptor);
 
             downloadFileIfMissing(downloadUrl, archive, this.userName, this.password);
 
@@ -213,20 +231,21 @@ public class NodeInstaller {
             File nodeBinary = new File(tmpDirectory, longNodeFilename + File.separator + "node.exe");
             if (!nodeBinary.exists()) {
                 throw new FileNotFoundException(
-                    "Could not find the downloaded Node.js binary in " + nodeBinary);
+                        "Could not find the downloaded Node.js binary in " + nodeBinary);
             } else {
-                File destinationDirectory = getInstallDirectory();
+                File destinationDirectory = getInstallDirectory(config);
 
                 File destination = new File(destinationDirectory, "node.exe");
                 this.logger.info("Copying node binary from {} to {}", nodeBinary, destination);
                 if (!nodeBinary.renameTo(destination)) {
                     throw new InstallationException("Could not install Node: Was not allowed to rename "
-                        + nodeBinary + " to " + destination);
+                            + nodeBinary
+                            + " to "
+                            + destination);
                 }
 
                 if ("provided".equals(this.npmVersion)) {
-                    File tmpNodeModulesDir =
-                        new File(tmpDirectory, longNodeFilename + File.separator + "node_modules");
+                    File tmpNodeModulesDir = new File(tmpDirectory, longNodeFilename + File.separator + "node_modules");
                     File nodeModulesDirectory = new File(destinationDirectory, "node_modules");
                     FileUtils.copyDirectory(tmpNodeModulesDir, nodeModulesDirectory);
                 }
@@ -244,20 +263,19 @@ public class NodeInstaller {
 
     }
 
-    private void installNodeForWindows() throws InstallationException {
+    private void installNodeForWindows(InstallConfig execConfig, InstallConfig config) throws InstallationException {
         final String downloadUrl = this.nodeDownloadRoot
-            + this.config.getPlatform().getNodeDownloadFilename(this.nodeVersion, false);
+                + config.getPlatform().getNodeDownloadFilename(this.nodeVersion, false);
         try {
-            File destinationDirectory = getInstallDirectory();
+            File destinationDirectory = getInstallDirectory(config);
 
             File destination = new File(destinationDirectory, "node.exe");
 
-            String classifier = this.config.getPlatform().getNodeClassifier();
+            String classifier = config.getPlatform().getNodeClassifier();
 
-            CacheDescriptor cacheDescriptor =
-                new CacheDescriptor("node", this.nodeVersion, classifier, "exe");
+            CacheDescriptor cacheDescriptor = new CacheDescriptor("node", this.nodeVersion, classifier, "exe");
 
-            File binary = this.config.getCacheResolver().resolve(cacheDescriptor);
+            File binary = config.getCacheResolver().resolve(cacheDescriptor);
 
             downloadFileIfMissing(downloadUrl, binary, this.userName, this.password);
 
@@ -272,8 +290,8 @@ public class NodeInstaller {
         }
     }
 
-    private File getTempDirectory() {
-        File tmpDirectory = new File(getInstallDirectory(), "tmp");
+    private File getTempDirectory(InstallConfig execConfig) {
+        File tmpDirectory = new File(getInstallDirectory(execConfig), "tmp");
         if (!tmpDirectory.exists()) {
             this.logger.debug("Creating temporary directory {}", tmpDirectory);
             tmpDirectory.mkdirs();
@@ -281,8 +299,8 @@ public class NodeInstaller {
         return tmpDirectory;
     }
 
-    private File getInstallDirectory() {
-        File installDirectory = new File(this.config.getInstallDirectory(), INSTALL_PATH);
+    private File getInstallDirectory(InstallConfig config) {
+        File installDirectory = new File(config.getInstallDirectory(), INSTALL_PATH);
         if (!installDirectory.exists()) {
             this.logger.debug("Creating install directory {}", installDirectory);
             installDirectory.mkdirs();
@@ -303,14 +321,14 @@ public class NodeInstaller {
     }
 
     private void downloadFileIfMissing(String downloadUrl, File destination, String userName, String password)
-        throws DownloadException {
+                                                                                                               throws DownloadException {
         if (!destination.exists()) {
             downloadFile(downloadUrl, destination, userName, password);
         }
     }
 
     private void downloadFile(String downloadUrl, File destination, String userName, String password)
-        throws DownloadException {
+                                                                                                      throws DownloadException {
         this.logger.info("Downloading {} to {}", downloadUrl, destination);
         this.fileDownloader.download(downloadUrl, destination.getPath(), userName, password);
     }